Ai Upscaler
This is a script i'm writing based on the super-image library. This script takes an input directory of images and uses the super-images' ai model to upscale the images x2. It works by doing small patches of the large images piece by piece and sews the results back together to create the larger image. This allows the user to work with images that their processor normally couldn't handle.